feat: WIP android refactor
