+++ backends/gdx-backend-robovm/src/com/badlogic/gdx/backends/iosrobovm/DefaultIOSInput.java
@@ -178,7 +178,7 @@ public class DefaultIOSInput extends AbstractInput implements IOSInput {
 	}
 
 	protected IOSNativeInput createNativeInput (NativeInputConfiguration configuration) {
-		return new IOSNativeInput(app, this, configuration);
+		return new IOSNativeInput(app, configuration);
 	}
 
 	@Override
@@ -515,7 +515,7 @@ public class DefaultIOSInput extends AbstractInput implements IOSInput {
 		this.observer = observer;
 	}
 
-	protected UIKeyboardType getIosInputType (OnscreenKeyboardType type) {
+	protected static UIKeyboardType getIosInputType (OnscreenKeyboardType type) {
 		UIKeyboardType preferredInputType;
 		switch (type) {
 		case NumberPad:
